Move translation tests
